+++ app/static/src/js/main.js
@@ -8,6 +8,10 @@ import { jobStatus } from "./components/job-status.js";
 import { copyButton } from "./components/copy.js";
 import { mathCaptcha } from "./components/captcha-math.js";
 import { toast } from "./components/toast.js";
+import { chartTimeline } from "./components/chart-timeline.js";
+import { initCapWidget } from "./components/cap-widget-setup.js";
+
+initCapWidget();
 
 Alpine.plugin(collapse);
 Alpine.plugin(focus);
@@ -18,6 +22,7 @@ Alpine.data("jobStatus", jobStatus);
 Alpine.data("copyButton", copyButton);
 Alpine.data("mathCaptcha", mathCaptcha);
 Alpine.data("toast", toast);
+Alpine.data("chartTimeline", chartTimeline);
 
 window.Alpine = Alpine;
 Alpine.start();